backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.4
scikit-learn==1.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.24.3 